frontend: Fixed broken npm run build
