+++ frontend/src/pages/404/index.js
@@ -1,4 +1,3 @@
-//export { default } from "./404";
 import React from "react";
 import Layout from "~/components/layout";
 import SEO from "~/components/seo";
